+++ src/com/carolinarollergirls/scoreboard/ScoreBoardManager.java
@@ -18,14 +18,14 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.concurrent.ConcurrentHashMap;
 
-import com.carolinarollergirls.scoreboard.base.ScoreBoard;
 import com.carolinarollergirls.scoreboard.defaults.DefaultScoreBoardModel;
 import com.carolinarollergirls.scoreboard.jetty.JettyServletScoreBoardController;
 import com.carolinarollergirls.scoreboard.json.JSONStateManager;
 import com.carolinarollergirls.scoreboard.json.JSONStateSnapshotter;
 import com.carolinarollergirls.scoreboard.json.ScoreBoardJSONListener;
 import com.carolinarollergirls.scoreboard.model.ScoreBoardModel;
-import com.carolinarollergirls.scoreboard.utils.ScoreBoardMetricsCollector;
+import com.carolinarollergirls.scoreboard.view.ScoreBoard;
+import com.carolinarollergirls.scoreboard.viewer.ScoreBoardMetricsCollector;
 import com.carolinarollergirls.scoreboard.viewer.TwitterViewer;
 
 public class ScoreBoardManager {
