+++ src/com/carolinarollergirls/scoreboard/jetty/XmlScoreBoardServlet.java
@@ -19,7 +19,7 @@ import org.jdom.JDOMException;
 import org.jdom.xpath.XPath;
 
 import com.carolinarollergirls.scoreboard.ScoreBoardManager;
-import com.carolinarollergirls.scoreboard.base.ScoreBoard;
+import com.carolinarollergirls.scoreboard.view.ScoreBoard;
 
 public class XmlScoreBoardServlet extends AbstractXmlServlet
 {
