+++ src/com/carolinarollergirls/scoreboard/viewer/ScoreBoardMetricsCollector.java
@@ -1,4 +1,4 @@
-package com.carolinarollergirls.scoreboard.utils;
+package com.carolinarollergirls.scoreboard.viewer;
 /**
  * Copyright (C) 2008-2012 Mr Temper <MrTemper@CarolinaRollergirls.com>
  *
@@ -15,9 +15,9 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import com.carolinarollergirls.scoreboard.base.Clock;
-import com.carolinarollergirls.scoreboard.base.ScoreBoard;
-import com.carolinarollergirls.scoreboard.base.Team;
+import com.carolinarollergirls.scoreboard.view.Clock;
+import com.carolinarollergirls.scoreboard.view.ScoreBoard;
+import com.carolinarollergirls.scoreboard.view.Team;
 
 
 public class ScoreBoardMetricsCollector extends Collector {
